+++ VTS3/BACnetLibrary/BACnet.hpp
@@ -495,7 +495,7 @@ class BACnetDate : public BACnetEncodeable {
 		void Decode( const char *dec );
 
 		// Convert to integer time in 100 nsec ticks since 1 Jan 1600 (FILETIME)
-		long long AsInt(void) const;
+		LONGLONG AsInt(void) const;
 
 		virtual int DataType(void);
 		virtual BACnetEncodeable * clone(void);
@@ -603,7 +603,7 @@ class BACnetDateRange : public BACnetEncodeable
 		bool operator <( const BACnetDateRange &arg );
 		bool operator >( const BACnetDateRange &arg );
 
-		long long GetSpan(void) const;
+		LONGLONG GetSpan(void) const;
 
 		virtual int DataType(void);
 		virtual BACnetEncodeable * clone(void);
